schnet_9m/requirements.txt
awscli
jsonargparse[all]
jupyter
natsort
pandas
periodictable
py3Dmol
pyarrow
rdkit
seaborn
torchinfo
torchmetrics
wandb
